Подправила код. Убрала поле с добавлением id

+++ front/src/features/comments/Comments.tsx
@@ -18,8 +18,11 @@ const Comments = () => {
   }, [dispatch]);
 
   const onDelete = async (id: string) => {
-    await dispatch(deleteComment(id));
-    await dispatch(fetchComments());
+    const confirmed = window.confirm('Are you sure you want to delete this comment?');
+    if (confirmed) {
+      await dispatch(deleteComment(id));
+      await dispatch(fetchComments());
+    }
   };
 
   return (
